[CLEANUP] Refactoring common JobMeta and TransMeta code up into AbstractMeta
